- replaced all of the old way to display vehicle, with the new database method

+++ app/src/main/java/model/DTO/NSchedulesMapMarkers.kt
@@ -4,7 +4,7 @@ class NSchedulesMapMarkers {
     companion object {
         fun retrieveVehicles(lineId: Int, serviceIds: List<Int>, callback: (List<NSchedulesMapMarker>) -> Unit) {
             Services.getServicesByLine(lineId) { services ->
-                val filteredServices = services.filter { serviceIds.contains(it.vehicle.id) }
+                val filteredServices = services.filter { serviceIds.contains(it.vehicle.id.toInt()) }
                 callback(filteredServices.map { NSchedulesMapMarker(NSchedulesMapMarkerType.VEHICLE, null, it) })
             }
         }
